Rework adapter binding logic.
